updated experiment 1 for optimizations

+++ prototypes/svae_proto/exp_001_vocab_trigram_recall/cfg.py
@@ -157,7 +157,7 @@ CFG_FRESNEL_128: Dict[str, Any] = dict(
     linear_readout=False,
     svd_mode='none',
     svd_method='gram_eigh',                 # auto routes D=16 to gram_eigh fp64
-    svd_compute_dtype='fp32',          # bug with triton needs to be reset to fp32 for now
+    svd_compute_dtype='fp64',          # bug with triton needs to be reset to fp32 for now
 
     # Training — fresnel_small's lr=1e-4 batch=128 30 epochs is enough
     # at full corpus given 17M params. ds_size scaled down because each
